% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gozer.R
\name{gozer}
\alias{gozer}
\title{Do something after an R object is destroyed (i.e. garbage collected)}
\usage{
gozer(obj, destructor)
}
\arguments{
\item{obj}{any R object}

\item{destructor}{string to print or function to call.  Note: function
takes a single argument which should be discarded}
}
\value{
the object with an an added attribute.  A 'destructor' artefact which
        will be printed/called after the object is destroyed.
}
\description{
Do something after an R object is destroyed (i.e. garbage collected)
}
